a5 opt

+++ src/main/scala/ca/uwaterloo/cs451/a5/Q2.scala
@@ -58,7 +58,7 @@ object Q2 {
         .filter(_._2._1.nonEmpty)
         .sortByKey()
         .take(20)
-        .map(p => (p._2._2.head, p._1.toLong))
+        .map(p => (p._2._2.head, p._1))
         .foreach(println)
     } else if (args.parquet()) {
       val sparkSession = SparkSession.builder.getOrCreate
